requirements.txt
pandas==2.2.3
pymysql==1.1.1
python-dotenv==1.0.1
SQLAlchemy==1.4.52
streamlit==1.39.0
